aura_api/requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
ed25519==1.5
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
PyJWT==2.6.0
python-arango==7.5.2
requests==2.28.1
requests-toolbelt==0.10.1
urllib3==1.26.12
Werkzeug==2.2.2
